Add commands, Add tests, Refactor
- Add file-exists command
- Add file-upload-single command
- Add file-download-single command
- Add device-list command
- Add device-update command
- Add tests
- Change global broker port from 8883 to 443
- Modify proces_response to set the last_response

This commit basically adds all the components to execute firmware updates
using mqtt-rpc-client-qt.
A device updating flow would usually look like this:
1. ping the Venus device to check if it is online
2. get the list of devices connected to the Venus device using device-list
3. (optional) check if the new firmware for a device is present on the
   Venus device (file-exists)
4. upload the firmware file to the Venus device using file-upload-single
5. start the actual firmware update by calling device-update

+++ src/util.cpp
@@ -7,7 +7,7 @@ QString token_urlsafe(qint32 length) {
 	QString randomString(length, ' ');
 	for(int i=0; i<length; ++i)
 	{
-		int index = qrand() % sizeof(possibleCharacters);
+		int index = static_cast<unsigned long>(qrand()) % sizeof(possibleCharacters);
 		randomString[i] = possibleCharacters[index];
 	}
 	return randomString;
